리뷰 업데이트 불러오기 진행 완료

+++ app/src/main/java/com/makeus/android/moari/interfaces/DialogRatingInterface.java
@@ -1,5 +1,5 @@
 package com.makeus.android.moari.interfaces;
 
 public interface DialogRatingInterface {
-    void rating(double tmp);
+    void rating(float tmp);
 }
